requirements.txt
Django==6.0.3
numpy>=2.0,<3.0
onnxruntime>=1.22,<2.0
Pillow>=11.0,<12.0
redis>=5.0,<6.0
aws-embedded-metrics>=3.0,<4.0
python-json-logger>=3.0,<4.0
